requirements.txt
Flask==3.0.0
pandas==2.2.0
numpy==1.26.2
Werkzeug==3.0.1 
requests~=2.32.4
undetected-chromedriver~=3.5.5
selenium~=4.34.0
beautifulsoup4~=4.13.4
langchain~=0.3.26
python-dotenv~=1.1.0
langchain-core~=0.3.66
firecrawl-py==2.15.0
pydantic~=2.11.7
pydantic-settings>=2.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
redis>=5.0.0
spacy>=3.7.0
typer>=0.9.0
rich>=13.0.0
alembic>=1.13.0
reportlab>=4.0.0
openai>=1.0.0
